fix(ProgressStep, StructuredList): fix exports
- make default export for each sub-component
- add hook for error messages
