+++ packages/react/src/components/ProgressIndicator/index.js
@@ -6,4 +6,5 @@
  */
 
 export * from './ProgressIndicator.Skeleton';
-export * from './ProgressIndicator';
+export { default as ProgressIndicator } from './ProgressIndicator';
+export { default as ProgressStep } from './ProgressStep';
